:sparkles: feat: Update Professional Maintenance With Toast and Modal

+++ src/app/modules/maintenance/pages/professionals-table-page/professionals-table-page.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Page } from 'src/app/core/models/page';
 import { Professional } from 'src/app/core/models/professional';
 import { ProfessionalService } from 'src/app/core/services/professional.service';
+import { ToastService } from 'src/app/core/services/toast.service';
+import { ModalComponent } from 'src/app/shared/components/modal/modal.component';
 
 @Component({
   selector: 'app-professionals-table-page',
@@ -10,14 +12,14 @@ import { ProfessionalService } from 'src/app/core/services/professional.service'
 })
 export class ProfessionalsTablePageComponent implements OnInit {
 
-  constructor(private professionalService: ProfessionalService) { }
+  constructor(private professionalService: ProfessionalService, private toastService: ToastService) { }
 
   professionalPage: Page<Professional> = {} as Page<Professional>;
   page = 1;
 
-  professionals: Professional[] = [];
-
   filter: string = "";
+
+  selectedProfessional !: Professional;
   
   ngOnInit(): void {
     this.loadProfessionals();
@@ -40,15 +42,20 @@ export class ProfessionalsTablePageComponent implements OnInit {
     this.loadProfessionals();
   }
 
-  deleteProfessional(professional: Professional) {
-    this.professionalService.deleteProfessional(professional).subscribe({
-      next: () => {
-        console.log("Excluido " + professional.id);
-        this.loadProfessionals();
-      },
-      error: () => {
-        console.log("Erro ao excluir");
+  deleteProfessional(professional: Professional, modalConfirm: ModalComponent) {
+    this.selectedProfessional = professional;
+    modalConfirm.open().then(confirm => {
+      if (confirm) {
+        this.professionalService.deleteProfessional(professional).subscribe({
+          next: () => {
+            this.toastService.show("Profissional Removido com sucesso!", { classname: "bg-success text-light" });
+            this.loadProfessionals();
+          },
+          error: () => {
+            this.toastService.show("Erro ao Remover o profissional!", { classname: "bg-danger text-light" })
+          }
+        });
       }
-    });
+    })
   }
 }
